requirements.txt
groq
openai
pinecone-client
sentence-transformers
pypdf
python-dotenv
numpy
torch
tqdm
pandas
scikit-learn
fastapi
uvicorn